+++ src/helpers/ai/ai.ts
@@ -3,12 +3,11 @@
  * @module helpers/ai/ai
  */
 
-import type { Knex } from 'knex';
-
-import { Catalog } from '../../models/catalog/catalog';
-import type { Request } from '../../types';
-
 const { config } = require(`${process.cwd()}/config`);
+
+interface VisionResult {
+  response: string;
+}
 
 /**
  * Embed
@@ -17,7 +16,7 @@ const { config } = require(`${process.cwd()}/config`);
  * @returns {string} Vector of input value
  */
 export async function embed(input: string): Promise<string> {
-  const response = await fetch(config.ai.models.embed.api, {
+  const response = await fetch(config.ai?.models?.embed?.api, {
     method: 'POST',
     headers: {
       'Content-Type': 'application/json',
@@ -42,14 +41,36 @@ export async function generate(
   prompt: string,
   context: string,
 ): Promise<string> {
-  const response = await fetch(config.ai.models.generate.api, {
+  const response = await fetch(config.ai?.models?.llm?.api, {
     method: 'POST',
     headers: {
       'Content-Type': 'application/json',
     },
     body: JSON.stringify({
-      model: config.ai.models.generate.name,
+      model: config.ai?.models?.llm?.name,
       prompt: `Query: ${prompt}\nContext: ${context}\nPlease provide an answer and use the given context if you can't find the answer.`,
+      stream: false,
+    }),
+  });
+  return await response.json();
+}
+
+/**
+ * Vision
+ * @method vision
+ * @param {string} data Image data to be processed
+ * @returns {VisionResult} response
+ */
+export async function vision(data: string): Promise<VisionResult> {
+  const response = await fetch(config.ai?.models?.vision?.api, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify({
+      model: config.ai?.models?.vision?.name,
+      prompt: 'What is in this picture?',
+      images: [data],
       stream: false,
     }),
   });