+++ CHANGELOG.md
@@ -55,6 +55,7 @@
 - Add similarity property to @related endpoint @robgietema
 - Fetch indexes from db instead of profiles @robgietema
 - Add search based on embedding model @robgietema
+- Add image recognition and indexing using a vision model @robgietema
 
 ### Bugfix